# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 디지털 디톡스 앱 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ServerTime {
  timestamp: string;
  isSunday: boolean;
  dayOfWeek: number;
  daysUntilSunday: number;
}

export interface TimeSlot {
  index: number;
  time: string;
  isLocked: boolean;
  /**
   * @minimum 0
   * @maximum 6
   */
  dayOfWeek: number;
}

export type UpdateTimeSlotsRequestSlotsItem = {
  index: number;
  isLocked: boolean;
};

export interface UpdateTimeSlotsRequest {
  /**
   * Day of week to update (0=Sun, 1=Mon, ..., 6=Sat)
   * @minimum 0
   * @maximum 6
   */
  day: number;
  slots: UpdateTimeSlotsRequestSlotsItem[];
}

export interface App {
  id: number;
  name: string;
  category: string;
  blocked: boolean;
  icon?: string;
  packageName?: string;
  bundleId?: string;
  usageMinutes?: number;
}

export type UpdateBlockedAppsRequestAppsItem = {
  id: number;
  blocked: boolean;
};

export interface UpdateBlockedAppsRequest {
  apps: UpdateBlockedAppsRequestAppsItem[];
}

export type ChallengeTier = (typeof ChallengeTier)[keyof typeof ChallengeTier];

export const ChallengeTier = {
  beginner: "beginner",
  motivated: "motivated",
  focused: "focused",
  hardcore: "hardcore",
} as const;

export interface Challenge {
  id?: number;
  tier: ChallengeTier;
  tierName: string;
  depositAmount: number;
  currentDay?: number;
  totalDays: number;
  successDays?: number;
  failedDays?: number;
  totalParticipants?: number;
  currentSurvivors?: number;
  totalPool?: number;
  potentialReward?: number;
  feeRate?: number;
  maxDailyHours?: number;
  dailyLimitHours?: number;
  isActive: boolean;
  startDate?: string;
  endDate?: string;
}

export type JoinChallengeRequestTier =
  (typeof JoinChallengeRequestTier)[keyof typeof JoinChallengeRequestTier];

export const JoinChallengeRequestTier = {
  beginner: "beginner",
  motivated: "motivated",
  focused: "focused",
  hardcore: "hardcore",
} as const;

export interface JoinChallengeRequest {
  tier: JoinChallengeRequestTier;
  /**
   * 하루 최대 사용 가능 시간 (티어 상한 이하로 설정)
   * @minimum 1
   * @maximum 24
   */
  dailyLimitHours?: number;
}

export interface TierStat {
  tier: string;
  total: number;
  active: number;
  completed: number;
}

export interface ChallengeStats {
  tiers: TierStat[];
  totalActive: number;
  grandTotal: number;
}

export interface HeartbeatResponse {
  ok: boolean;
  timestamp: string;
}

export type ChallengeEventType =
  (typeof ChallengeEventType)[keyof typeof ChallengeEventType];

export const ChallengeEventType = {
  joined: "joined",
  quit: "quit",
  eliminated: "eliminated",
} as const;

export interface ChallengeEvent {
  id: number;
  type: ChallengeEventType;
  nickname: string;
  tier: string;
  tierName: string;
  createdAt: string;
}

export interface Dashboard {
  isLocked: boolean;
  /** Today's day of week in KST (0=Sun, 1=Mon ... 6=Sat) */
  todayDow: number;
  currentStreak: number;
  todayComplianceRate: number;
  totalDays: number;
  challenge?: Challenge;
  timeSlots: TimeSlot[];
  todayFocusMinutes?: number;
  weeklyFocusMinutes?: number[];
}

export interface Settings {
  nickname?: string;
  notificationsEnabled: boolean;
  lockReminderEnabled: boolean;
  challengeAlertEnabled: boolean;
  pushPermissionGranted?: boolean;
  deviceType?: string;
  timeslotsConfigured?: boolean;
  appsConfigured?: boolean;
}

export interface UpdateSettingsRequest {
  nickname?: string;
  notificationsEnabled?: boolean;
  lockReminderEnabled?: boolean;
  challengeAlertEnabled?: boolean;
}

export interface StartFocusTimerRequest {
  durationMinutes: number;
}

export interface FocusSession {
  id: number;
  durationMinutes: number;
  completedMinutes?: number;
  startedAt?: string;
  completedAt?: string;
  isCompleted: boolean;
}

export type UsageStatsByAppItem = {
  appName: string;
  minutes: number;
};

export interface UsageStats {
  date: string;
  totalMinutes: number;
  byApp?: UsageStatsByAppItem[];
}

export type UpdateUsageStatsRequestByAppItem = {
  appName: string;
  minutes: number;
};

export interface UpdateUsageStatsRequest {
  totalMinutes: number;
  byApp?: UpdateUsageStatsRequestByAppItem[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetTimeSlotsParams = {
  /**
   * Day of week (0=Sun, 1=Mon, ..., 6=Sat). Defaults to today.
   * @minimum 0
   * @maximum 6
   */
  day?: number;
};

export type QuitChallenge200 = {
  ok: boolean;
};

export type GetRecentChallengeEventsParams = {
  /**
   * ISO timestamp — return events after this time
   */
  since?: string;
};