# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 디지털 디톡스 앱 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /challenge:
    get:
      operationId: getChallenge
      tags: [challenge]
      summary: Get current challenge info
      responses:
        "200":
          description: Challenge info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
    post:
      operationId: joinChallenge
      tags: [challenge]
      summary: Join a challenge (Sunday only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinChallengeRequest"
      responses:
        "200":
          description: Joined challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "403":
          description: Not Sunday - settings locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /challenge/events/recent:
    get:
      operationId: getRecentChallengeEvents
      tags: [challenge]
      summary: Get recent challenge events for polling
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: ISO timestamp — return events after this time
      responses:
        "200":
          description: Recent events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChallengeEvent"
...
  /focus-timer:
    post:
      operationId: startFocusTimer
      tags: [stats]
      summary: Start a focus timer session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartFocusTimerRequest"
      responses:
        "200":
          description: Focus session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"
    get:
      operationId: getFocusSessions
      tags: [stats]
      summary: Get focus sessions history
      responses:
        "200":
          description: List of focus sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FocusSession"
  /focus-timer/{sessionId}/complete:
    post:
      operationId: completeFocusTimer
      tags: [stats]
      summary: Complete a focus timer session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"
...
components:
  schemas:
...
    Challenge:
      type: object
      properties:
        id:
          type: integer
        tier:
          type: string
          enum: [beginner, motivated, focused, hardcore]
        tierName:
          type: string
        depositAmount:
          type: integer
        currentDay:
          type: integer
        totalDays:
          type: integer
        successDays:
          type: integer
        failedDays:
          type: integer
        totalParticipants:
          type: integer
        currentSurvivors:
          type: integer
        totalPool:
          type: integer
        potentialReward:
          type: integer
        feeRate:
          type: number
        maxDailyHours:
          type: integer
        dailyLimitHours:
          type: integer
        isActive:
          type: boolean
        startDate:
          type: string
        endDate:
          type: string
      required: [tier, tierName, depositAmount, totalDays, isActive]
    JoinChallengeRequest:
      type: object
      properties:
        tier:
          type: string
          enum: [beginner, motivated, focused, hardcore]
        dailyLimitHours:
          type: integer
          minimum: 1
          maximum: 24
          description: 하루 최대 사용 가능 시간 (티어 상한 이하로 설정)
      required: [tier]
...
    ChallengeEvent:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [joined, quit, eliminated]
        nickname:
          type: string
        tier:
          type: string
        tierName:
          type: string
        createdAt:
          type: string
      required: [id, type, nickname, tier, tierName, createdAt]
...
    StartFocusTimerRequest:
      type: object
      properties:
        durationMinutes:
          type: integer
      required: [durationMinutes]
    FocusSession:
      type: object
      properties:
        id:
          type: integer
        durationMinutes:
          type: integer
        completedMinutes:
          type: integer
        startedAt:
          type: string
        completedAt:
          type: string
        isCompleted:
          type: boolean
      required: [id, durationMinutes, isCompleted]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]