# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 디지털 디톡스 앱 API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeTier } from "./challengeTier";

export interface Challenge {
  id?: number;
  tier: ChallengeTier;
  tierName: string;
  depositAmount: number;
  currentDay?: number;
  totalDays: number;
  successDays?: number;
  failedDays?: number;
  totalParticipants?: number;
  currentSurvivors?: number;
  totalPool?: number;
  potentialReward?: number;
  feeRate?: number;
  maxDailyHours?: number;
  dailyLimitHours?: number;
  isActive: boolean;
  startDate?: string;
  endDate?: string;
}